+++ sure-core/src/main/java/com/sure/tool/util/NumberUtil.java
@@ -545,5 +545,4 @@ public class NumberUtil {
 		return isInteger(str);
 	}
 
-
 }